+++ projects/igniteui-angular/src/lib/query-builder/query-builder.component.ts
@@ -7,7 +7,7 @@ import {
 import { FormsModule } from '@angular/forms';
 import { Subject } from 'rxjs';
 import { editor } from '@igniteui/material-icons-extended';
-import { IButtonGroupEventArgs, IgxButtonGroupComponent } from '../buttonGroup/buttonGroup.component';
+import { IButtonGroupAfterEventArgs, IgxButtonGroupComponent } from '../buttonGroup/buttonGroup.component';
 import { IgxChipComponent } from '../chips/chip.component';
 import { DisplayDensityBase, DisplayDensityToken, IDisplayDensityOptions } from '../core/density';
 import { IQueryBuilderResourceStrings, QueryBuilderResourceStringsEN } from '../core/i18n/query-builder-resources';
@@ -785,7 +785,7 @@ export class IgxQueryBuilderComponent extends DisplayDensityBase implements Afte
     /**
      * @hidden @internal
      */
-    public selectFilteringLogic(event: IButtonGroupEventArgs) {
+    public selectFilteringLogic(event: IButtonGroupAfterEventArgs) {
         this.contextualGroup.operator = event.index as FilteringLogic;
         this.commitOperandEdit();
     }