+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes for each version of this project will be documented in this 
 - We're working on reducing the library size
     - IgxRadioComponent has been reduced in half
     - IgxSwitchComponent has been reduced in half
+- `IgxButtonGroup`:
+    - Adding `selecting` and `deselecting` events that are fired before selecting or deselecting a button and they are also cancellable.
 
 ## 16.1.4
 ### New Features
@@ -45,7 +47,7 @@ All notable changes for each version of this project will be documented in this 
     @include theme();
     ```
 - `IgxButtonGroup`:
-    - Reverted cancellable on `selected` and `deselected` events (added in 15.1.0) as it was breaking firing order. 
+    - Reverted cancellable on `selected` and `deselected` events (added in 15.1.0) as it was breaking firing order.
 
 ## 16.1.0
 ### New Features